Remove DayTimeHelper and move logic to enum
Most of the time "Helper" Classes are bad design. In this case the DayTimeHelper contained model logic that should ideally belong to the DayTime model (enum) itself, for example creation logic when providing an hour as an Int. This seems to be an ideal case for an initializer in an enum, as it effectivly initializes a DayTime from a given Int. Through the "Helper" Design this logic is scattered accross different places.
Therefore I moved this logic to the enum itself, making it very clear to a potential user how to create a DayTime.

The same is true for some other functionality the DayTimeHelper provided.