Remove WallpaperGroup's load() in favour of lazy var
Removes the load() function in WallpaperGroup in favour of a lazily initialized wallpapers array. "load"-functions most of the time do not provide any real value an make it more difficult and error prone to use a class. Some user of WallpaperGroup might initialize an instance of the class and *not* know about the "load()" function - thus receiving an object that has not been initialized completly.
In this case a lazily initialized property seems to be perfect. The user does not have to know the inner workings of the class and can just use it. On the other hand, no work is done that is uneeded. The property will only be initialized if it is needed.