Remove unsafe force unwrapping
Force unwrapping is unsafe in this case, as a caller to this method might pass a NSURL object that has no associated path. This would then result in a crash due to a nil value. To fix this potential issue, I removed the force unwrapping and introduced a guard statement as a pre condition for the method.